v1.0

+++ scripts/autoupdate.py
@@ -7,8 +7,12 @@ import pexpect
 import psutil
 import os
 
+def log_print(string):
+    print(f"[{datetime.datetime.now().strftime('%Y-%m-%d_%H:%M:%S')}] {string}")
+
+
 class AutoUpdate:
-    def __init__(self, interval, git_address, temp_dir, password, executable_path):
+    def __init__(self, interval, ip, git_address, temp_dir, password, executable_path):
 
         self.interval = interval
 
@@ -16,12 +20,21 @@ class AutoUpdate:
         self.temp_dir = temp_dir
         self.password = password
 
+        self.ip = ip
+
         self.executable_path = executable_path
 
         self.log_txt = os.path.join(self.temp_dir, "updatelog.txt")
 
+    def fingerprint(self):
+        cmd = f"ssh-keyscan -t ed25519 {self.ip} >> ~/.ssh/known_hosts"
+        log_print(f"command: {cmd}")
+        subprocess.run(cmd, shell=True, executable="/bin/bash", capture_output=True, text=True, timeout=10)
+        log_print("fingerprint finished")
+
     def clone(self,git_address, temp_dir, password):
         cmd = f"cd {temp_dir} && git clone -b main {git_address}"
+        log_print(f"command: {cmd}")
 
         child = pexpect.spawn("/bin/bash", ["-c", cmd], timeout=60, encoding='utf-8')
 
@@ -30,29 +43,35 @@ class AutoUpdate:
             child.sendline(password)
             child.expect(pexpect.EOF, timeout=120)
         except pexpect.exceptions.TIMEOUT:
-            print("Timeout, password prompt not appeared or too slow")
+            log_print("Timeout, password prompt not appeared or too slow")
+            return False
         except pexpect.exceptions.EOF:
-            print("Process ended")
+            log_print("Process ended")
+            return False
         finally:
             child.close()
-        
-        if child.exitstatus != 0:
-            raise RuntimeError(f"git clone failed, exit code: {child.exitstatus}")
+
+        return True        
+        # if child.exitstatus != 0:
+        #     raise RuntimeError(f"git clone failed, exit code: {child.exitstatus}")
         
 
     def git_clone(self):
-        print("Cloning...")
+        log_print("Cloning...")
 
         if(os.path.exists(f"{self.temp_dir}/ai_algorithm_bin")):
             shutil.rmtree(f"{self.temp_dir}/ai_algorithm_bin")
-        self.clone(self.git_address, self.temp_dir, self.password)
+        if(not self.clone(self.git_address, self.temp_dir, self.password)):
+            log_print("try to clone project but failed")
+            return False
 
 
         with open(self.log_txt, "a") as f:
             f.write(f"{datetime.datetime.now().strftime('%Y-%m-%d_%H:%M:%S')} Clone finished\n")
         
-        print(f"{datetime.datetime.now().strftime('%Y-%m-%d_%H:%M:%S')} Clone finished\n")
+        log_print(f"{datetime.datetime.now().strftime('%Y-%m-%d_%H:%M:%S')} Clone finished")
         
+        return True
         
 
     def is_executable_running(self, exec_path):
@@ -60,7 +79,7 @@ class AutoUpdate:
         for proc in psutil.process_iter(['pid', 'name', 'exe']):
             try:
                 if proc.exe() == abs_path:
-                    print(f"running! PID: {proc.pid}, cmdline: {proc.cmdline()}")
+                    log_print(f"running! PID: {proc.pid}, cmdline: {proc.cmdline()}")
                     return True
             except (psutil.NoSuchProcess, psutil.AccessDenied, psutil.ZombieProcess):
                 pass
@@ -71,6 +90,9 @@ class AutoUpdate:
         if self.is_executable_running(self.executable_path):
             return
         else:
+            if not os.path.exists(f"{self.temp_dir}/ai_algorithm_bin"):
+                return
+
             shutil.rmtree("/root/workspace/install/")
             shutil.copytree(
                 f"{self.temp_dir}/ai_algorithm_bin", 
@@ -80,7 +102,7 @@ class AutoUpdate:
             with open(self.log_txt, "a") as f:
                 f.write(f"{datetime.datetime.now().strftime('%Y-%m-%d_%H:%M:%S')} update executable finished\n")
             
-            print(f"{datetime.datetime.now().strftime('%Y-%m-%d_%H:%M:%S')} update executable finished\n")
+            log_print(f"{datetime.datetime.now().strftime('%Y-%m-%d_%H:%M:%S')} update executable finished")
 
     def run_once(self):
 
@@ -104,8 +126,8 @@ class AutoUpdate:
             shell=True, executable="/bin/bash", capture_output=True, text=True, timeout=300
         )
 
-        self.git_clone()
-        self.install()
+        if(self.git_clone()):
+            self.install()
 
         result = subprocess.run(
             "systemctl restart perception.service",
@@ -119,11 +141,13 @@ class AutoUpdate:
 
 AutoUpdate_t = AutoUpdate(
     interval=1,
+    ip = "61.129.72.17",
     git_address="git@61.129.72.17:/home/git/ai_algorithm_bin.git", 
     temp_dir="/home/sunrise/Desktop", 
     password="passgit5646",
     executable_path="/root/workspace/install/perception/lib/perception/perception_node"
 )
+AutoUpdate_t.fingerprint()
 AutoUpdate_t.run()
 
 # now_str = datetime.datetime.now().strftime("%Y-%m-%d_%H:%M:%S")
